preliminary replacement for post-install scriptlet
Carved out all of the logic to detect setup mode and create new keys, the script is now paired down to only attempting to sign detected kernel images, efi files, and vmlinuz files in /boot/ and /boot/efi. Intent is for this file to go into /usr/lib/kernel/install.d/ and to automate doing this with a modification to nobara-secure-boot.sh.

Reason for replacing a system file: The script supplied by sbctl is general purpose and makes certain assumptions about where your bootable files are. Nobara does not adhere to the directory structure expected by sbctl, so their post-install scriptlet fails to sign new versions of the kernel as they're installed. Replacing the system file supplied by sbctl may need to be re-done periodically as sbctl updates.